feat(frontend): add FileService for file size formatting and update FileUploadComponent
